Extract DNS response truncation into its own module
~110 lines of EDNS0/c-ares juggling (parse OPT, drop additional+authority,
drop answers until under the limit, set TC) was buried as static functions
in the UDP listener — only reachable by booting a real UDP server.

Move to dns_truncate.{c,h} with one entry point: dns_truncate_for_udp().
The UDP listener's respond path collapses to validate-length,
truncate-for-fit, sendto. The module is now unit-testable with synthetic
byte buffers.
