+++ src/dns_listener_udp.c
@@ -1,4 +1,3 @@
-#include <ares.h>
 #include <arpa/inet.h>
 #include <errno.h>
 #include <stdint.h>
@@ -9,6 +8,7 @@
 
 #include "dns_common.h"
 #include "dns_listener_udp.h"
+#include "dns_truncate.h"
 #include "logging.h"
 
 typedef struct dns_listener_udp_s {
@@ -86,116 +86,6 @@ static void watcher_cb(struct ev_loop __attribute__((unused)) *loop,
   d->cb(d->cb_data, &d->base, (struct sockaddr*)&tmp_raddr, dns_req, (size_t)len);
 }
 
-static uint16_t get_edns_udp_size(const char *dns_req, const size_t dns_req_len) {
-  ares_dns_record_t *dnsrec = NULL;
-  ares_status_t parse_status = ares_dns_parse((const unsigned char *)dns_req, dns_req_len, 0, &dnsrec);
-  if (parse_status != ARES_SUCCESS) {
-    WLOG("Failed to parse DNS request: %s", ares_strerror((int)parse_status));
-    return DNS_SIZE_LIMIT;
-  }
-  const uint16_t tx_id = ares_dns_record_get_id(dnsrec);
-  uint16_t udp_size = 0;
-  const size_t record_count = ares_dns_record_rr_cnt(dnsrec, ARES_SECTION_ADDITIONAL);
-  for (size_t i = 0; i < record_count; ++i) {
-    const ares_dns_rr_t *rr = ares_dns_record_rr_get(dnsrec, ARES_SECTION_ADDITIONAL, i);
-    if (ares_dns_rr_get_type(rr) == ARES_REC_TYPE_OPT) {
-      udp_size = ares_dns_rr_get_u16(rr, ARES_RR_OPT_UDP_SIZE);
-      if (udp_size > 0) {
-        DLOG("%04hX: Found EDNS0 UDP buffer size: %u", tx_id, udp_size);
-      }
-      break;
-    }
-  }
-  ares_dns_record_destroy(dnsrec);
-  if (udp_size < DNS_SIZE_LIMIT) {
-    DLOG("%04hX: EDNS0 UDP buffer size %u overruled to %d", tx_id, udp_size, DNS_SIZE_LIMIT);
-    return DNS_SIZE_LIMIT;  // RFC6891 4.3 "Values lower than 512 MUST be treated as equal to 512."
-  }
-  return udp_size;
-}
-
-static void truncate_dns_response(char *buf, size_t *buflen, const uint16_t size_limit) {
-  const size_t old_size = *buflen;
-  buf[2] |= 0x02;  // anyway: set truncation flag
-
-  ares_dns_record_t *dnsrec = NULL;
-  ares_status_t status = ares_dns_parse((const unsigned char *)buf, *buflen, 0, &dnsrec);
-  if (status != ARES_SUCCESS) {
-    WLOG("Failed to parse DNS response: %s", ares_strerror((int)status));
-    return;
-  }
-  const uint16_t tx_id = ares_dns_record_get_id(dnsrec);
-
-  // NOTE: according to current c-ares implementation, removing first or last elements are the fastest!
-
-  // remove every additional and authority record
-  while (ares_dns_record_rr_cnt(dnsrec, ARES_SECTION_ADDITIONAL) > 0) {
-    status = ares_dns_record_rr_del(dnsrec, ARES_SECTION_ADDITIONAL, 0);
-    if (status != ARES_SUCCESS) {
-      WLOG("%04hX: Could not remove additional record: %s", tx_id, ares_strerror((int)status));
-    }
-  }
-  while (ares_dns_record_rr_cnt(dnsrec, ARES_SECTION_AUTHORITY) > 0) {
-    status = ares_dns_record_rr_del(dnsrec, ARES_SECTION_AUTHORITY, 0);
-    if (status != ARES_SUCCESS) {
-      WLOG("%04hX: Could not remove authority record: %s", tx_id, ares_strerror((int)status));
-    }
-  }
-
-  // rough estimate to reach size limit
-  size_t answers = ares_dns_record_rr_cnt(dnsrec, ARES_SECTION_ANSWER);
-  size_t answers_to_keep = ((size_limit - DNS_HEADER_LENGTH) * answers) / old_size;
-  answers_to_keep = answers_to_keep > 0 ? answers_to_keep : 1;  // try to keep 1 answer
-
-  // remove answer records until fit size limit or running out of answers
-  unsigned char *new_resp = NULL;
-  size_t new_resp_len = 0;
-  for (uint8_t g = 0; g < UINT8_MAX; ++g) {  // endless loop guard
-    status = ares_dns_write(dnsrec, &new_resp, &new_resp_len);
-    if (status != ARES_SUCCESS) {
-      WLOG("%04hX: Failed to create truncated DNS response: %s", tx_id, ares_strerror((int)status));
-      new_resp = NULL;  // just to be sure
-      break;
-    }
-    if (new_resp_len < size_limit || answers == 0) {
-      break;
-    }
-    if (new_resp_len >= old_size) {
-      WLOG("%04hX: Truncated DNS response size larger or equal to original: %u >= %u",
-           tx_id, new_resp_len, old_size);  // impossible?
-    }
-    ares_free_string(new_resp);
-    new_resp = NULL;
-
-    DLOG("%04hX: DNS response size truncated from %u to %u but to keep %u limit reducing answers from %u to %u",
-         tx_id, old_size, new_resp_len, size_limit, answers, answers_to_keep);
-
-    while (answers > answers_to_keep) {
-      status = ares_dns_record_rr_del(dnsrec, ARES_SECTION_ANSWER, answers - 1);
-      if (status != ARES_SUCCESS) {
-        WLOG("%04hX: Could not remove answer record: %s", tx_id, ares_strerror((int)status));
-        break;
-      }
-      --answers;
-    }
-    answers = ares_dns_record_rr_cnt(dnsrec, ARES_SECTION_ANSWER);  // update to be sure!
-    answers_to_keep /= 2;
-  }
-  ares_dns_record_destroy(dnsrec);
-
-  if (new_resp != NULL) {
-    if (new_resp_len < old_size) {
-      memcpy(buf, new_resp, new_resp_len);
-      *buflen = new_resp_len;
-      buf[2] |= 0x02;  // set truncation flag
-      ILOG("%04hX: DNS response size truncated from %u to %u to keep %u limit",
-           tx_id, old_size, new_resp_len, size_limit);
-    }
-    ares_free_string(new_resp);
-    new_resp = NULL;
-  }
-}
-
 static void udp_respond(dns_listener_t *self, struct sockaddr *raddr,
                         const char *dns_req, size_t dns_req_len,
                         char *dns_resp, size_t dns_resp_len) {
@@ -205,16 +95,7 @@ static void udp_respond(dns_listener_t *self, struct sockaddr *raddr,
     WLOG("Malformed response received, invalid length: %u", dns_resp_len);
     return;
   }
-  if (dns_resp_len > DNS_SIZE_LIMIT) {
-    const uint16_t udp_size = get_edns_udp_size(dns_req, dns_req_len);
-    if (dns_resp_len > udp_size) {
-      truncate_dns_response(dns_resp, &dns_resp_len, udp_size);
-    } else {
-      uint16_t tx_id = ntohs(*((uint16_t*)dns_req));
-      DLOG("%04hX: DNS response size %u larger than %d but EDNS0 UDP buffer size %u allows it",
-           tx_id, dns_resp_len, DNS_SIZE_LIMIT, udp_size);
-    }
-  }
+  dns_truncate_for_udp(dns_req, dns_req_len, dns_resp, &dns_resp_len);
 
   ssize_t len = sendto(d->sock, dns_resp, dns_resp_len, 0, raddr, d->addrlen);
   if (len == -1) {
